backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-genai==0.2.2
urllib3==2.1.0
python-dotenv==1.0.0
